+++ OneBlink.SDK/models/FormsAppStylesBase.cs
@@ -1,14 +1,10 @@
-using System.Collections.Generic;
-
 namespace OneBlink.SDK.Model
 {
-    public class FormsAppStyles
+    public class FormsAppStylesBase
     {
         public string foregroundColour { get; set; }
         public string highlightColour { get; set; }
         public string contrastColour { get; set; }
-        public string logoUrl { get; set; }
         public string customCss { get; set; }
-        public List<FormsAppMenuItems> menuItems { get; set; }
     }
 }