+++ OneBlink.SDK/models/FormsListStyles.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+
+namespace OneBlink.SDK.Model
+{
+    public class FormsListStyles : FormsAppStylesBase
+    {
+        public string logoUrl {get;set;}
+        public List<FormsAppMenuItems> menuItems {get;set;}
+    }
+}