+++ OneBlink.SDK/models/VolunteersFormsApp.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+
+namespace OneBlink.SDK.Model
+{
+    public class VolunteersFormsApp : FormsAppBase
+    {
+        public FormsAppStylesBase styles {get;set;}
+        public List<VolunteersFormsAppCategories> categories {get;set;}
+        public string waiverUrl {get;set;}
+    }
+}