+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 - `GenerateSubmissionDataUrl` to `FormsClient`
 - `GetMyFormsApp`, `CreateUser` and `DeleteUser` to `FormsAppClient`
+- `Get` to `OrganisationsClient`
+
+### Changed
+
+- Updated `FormsApp` to `ApprovalsFormsApp`, `FormsListFormsApp` and `VolunteersFormsApp`
 
 ## 1.17.0 (2021-03-23)